% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EQ5D2Utility.R
\name{Cad_EQ.5D_Utility}
\alias{Cad_EQ.5D_Utility}
\title{EQ5D2Utility}
\usage{
Cad_EQ.5D_Utility(x)
}
\arguments{
\item{x}{The input is a numeric vector of length 5}
}
\value{
Utility Value based on Canadian Value Set of EQ5D Index Scores
}
\description{
EQ5D2Utility
}
